+++ merkle/compact/tree.go
@@ -46,9 +46,9 @@ type Tree struct {
 //
 // The expectedRoot is the known-good tree root of the tree at the specified
 // size, and is used to verify the initial state.
-func NewTreeWithState(hasher hashers.LogHasher, size int64, hashes [][]byte, expectedRoot []byte) (*Tree, error) {
+func NewTreeWithState(hasher hashers.LogHasher, size uint64, hashes [][]byte, expectedRoot []byte) (*Tree, error) {
 	fact := RangeFactory{Hash: hasher.HashChildren}
-	rng, err := fact.NewRange(0, uint64(size), hashes)
+	rng, err := fact.NewRange(0, size, hashes)
 	if err != nil {
 		return nil, err
 	}
@@ -142,8 +142,8 @@ func (t *Tree) AppendLeafHash(leafHash []byte, visit VisitFn) error {
 }
 
 // Size returns the current size of the tree.
-func (t *Tree) Size() int64 {
-	return int64(t.rng.End())
+func (t *Tree) Size() uint64 {
+	return t.rng.End()
 }
 
 // hashes returns the set of node hashes that comprise the compact
